Genetic works with new elements, but does not give as good solution as greedy
Now Genetic gives a matching where the new elements are included. However, Greedy Algorithm always better! Not good! I believe this is because the initial population is only a small, small fraction of all possibilities, as the population consistst of random bits of 1 and 0. Next we should try to include a percentage of all the ALLOWED POSSIBILITIES, that Lars is now creating when developing Brute force